New. EmailEncoder. Shortcode to exclude data from encoding. Lot of refactoring.
